+++ persist_config.h
@@ -9,8 +9,8 @@ extern void        init_persistent();
 extern const char* persistent_get_name();
 extern void        persistent_set_name(const char * name[32]);
 
-extern bool        persistent_get_cal(unsigned adc, int32_t * scale, int32_t * offset);
-extern bool        persistent_set_cal(unsigned adc, int32_t scale, int32_t offset);
+extern bool        persistent_get_cal(unsigned adc, uint64_t * scale, uint64_t * offset);
+extern bool        persistent_set_cal(unsigned adc, uint64_t scale, uint64_t offset);
 
 
 #endif //__PERSISTENT_CONIFG__
